requirements.txt
asgiref==3.8.1 ; python_version >= "3.11" and python_version < "4.0"
django-cors-headers==4.3.1 ; python_version >= "3.11" and python_version < "4.0"
django==5.0.6 ; python_version >= "3.11" and python_version < "4.0"
djangorestframework==3.15.1 ; python_version >= "3.11" and python_version < "4.0"
drf-yasg==1.21.7 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==22.0.0 ; python_version >= "3.11" and python_version < "4.0"
inflection==0.5.1 ; python_version >= "3.11" and python_version < "4.0"
mysqlclient==2.2.4 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.0 ; python_version >= "3.11" and python_version < "4.0"
psycopg2-binary==2.9.9 ; python_version >= "3.11" and python_version < "4.0"
python-decouple==3.8 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.1 ; python_version >= "3.11" and python_version < "4.0"
pyyaml==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
sqlparse==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.1 ; python_version >= "3.11" and python_version < "4.0" and sys_platform == "win32"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
whitenoise==6.6.0 ; python_version >= "3.11" and python_version < "4.0"
